pyproject.toml
[tool.poetry]
name = "my-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
flask-cors = "^5.0.0"
requests = "^2.32.3"
pdfplumber = "^0.11.4"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
